Forced user to enter a valid input for restart.
The program was stopping when the user inputted something other than Y or N when asked to enter one of them in order to continue or stop the program. I added a while loop to solve the problem. Also changed the back color of the text "YOU WON!" to green to make it easier to catch for the user.